pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatbmw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.57.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["chatbmw*"]


[dependency-groups]
dev = [
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "huggingface-hub>=0.36.0",
    "jupyter>=1.1.1",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "notebook>=7.5.1",
    "pyyaml>=6.0.3",
    "rouge-score>=0.1.2",
    "trl>=0.24.0",
    "unsloth>=2026.1.2",
    "wandb>=0.23.1",
    "weave>=0.52.22",
    "tensorboard>=2.20.0",
    "bert-score>=0.3.13",
    "sacrebleu>=2.5.1",
]
